fix(cli-app-scripts): use the same build dir for modes (#302)
This removes the devOut path and usage of it, instead opting for
building the development build and production build into the same
folder: ./build.

For libraries in monorepo situations this is important as they will look
up the main or module key in the package.json file, which is usually set
to e.g. ./build/{cjs,es}/lib.js when using the cli-app-scripts to build
the lib.