fix(shell): create new stacking context for roots
This adds two rules for the application root and portal root divs that
creates a new stacking context for each, this allows an app the freedom
to do pretty much whatever it wants with z-indexes, without affecting UI
components that render in a Portal.

It also allows other 3rd party libraries that rely on appending elements
to the DOM and modifying z-indexes to work better with our platform.

The isolation property has good support and should align nicely with our
supported browsers[1].

[1] https://caniuse.com/mdn-css_properties_isolation