fix(app-icon): replace with placeholder icon
Change the default app icon to be an actual placeholder icon instead of
the DHIS2 logo.

The DHIS2 logo is part of the brand and we should not make everyone
potentially violate the license of how the DHIS2 brand can be used
without consent.

The solution is to replace the DHIS2 logo as the placeholder icon with a
generic placeholder icon.